requirements.txt
pymisp
python-dotenv
pandas
